pyproject.toml
[project]
name = "trial-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.1",
    "groq<1",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "notion-client>=3.0.0",
    "python-dotenv>=1.2.2",
]
